requirements.txt
Django==1.11.10
django-ckeditor==5.4.0
django-js-asset==1.0.0
django-mptt==0.9.0
django-nested-inline==0.3.7
django-widget-tweaks==1.4.1
Pillow==5.0.0
psycopg2==2.7.4
pytz==2018.3
